requirements.txt
google-api-python-client==2.126.0
python-dotenv==1.0.1
pandas==2.2.3
langdetect==1.0.9
vaderSentiment==3.3.2
matplotlib==3.10.0
streamlit==1.41.1
